+++ pages/LoadingPage/LoadingScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const LoadingScreen = () => {
+
+  return (
+    <View></View>
+  );
+};
+
+
+
+export default LoadingScreen;
