+++ pages/UserPage/FindUserIdScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const FindUserIdScreen = () => {
+
+  return (
+    <View></View>
+  );
+};
+
+
+
+export default FindUserIdScreen;
