+++ pages/UserPage/ForgotNameScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const ForgotNameScreen = () => {
+
+  return (
+    <View></View>
+  );
+};
+
+
+
+export default ForgotNameScreen;
