+++ pages/UserPage/ForgotPassScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const ForgotPassScreen = () => {
+
+  return (
+    <View></View>
+  );
+};
+
+
+
+export default ForgotPassScreen;
