+++ pages/UserPage/LoginScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const LoginScreen = () => {
+
+  return (
+    <View></View>
+  );
+};
+
+
+
+export default LoginScreen;
