+++ pages/UserPage/ResetPassScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const ResetPassScreen = () => {
+
+  return (
+    <View></View>
+  );
+};
+
+
+
+export default ResetPassScreen;
