+++ pages/UserPage/SettingScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View } from 'react-native';
+
+const SettingScreen = () => {
+
+  return (
+    <View></View>
+  );
+};
+
+
+
+export default SettingScreen;
